+++ src/components/FlexibleForm/ControlDecorators/partials/constants.js
@@ -1,0 +1,10 @@
+export const ACCEPTED_DATE_VALUES = {
+  TODAY: {
+    value: 'TODAY',
+    labelId: 'ui-data-import.settings.mappingProfiles.map.wrapper.acceptedValues.today',
+  },
+  CHOOSE_DATE: {
+    value: 'CHOOSE_DATE',
+    labelId: 'ui-data-import.settings.mappingProfiles.map.wrapper.acceptedValues.chooseDate',
+  },
+};